+++ tensorlayer/layers/recurrent.py
@@ -560,7 +560,7 @@ class BasicConvLSTMCell(ConvRNNCell):
     @property
     def state_size(self):
         """State size of the LSTMStateTuple."""
-        return LSTMStateTuple(self._num_units, self._num_units) if self._state_is_tuple else (2 * self._num_units)
+        return (LSTMStateTuple(self._num_units, self._num_units) if self._state_is_tuple else 2 * self._num_units)
 
     @property
     def output_size(self):
